Reset Datatypes for FFT

+++ src/fft.hpp
@@ -6,8 +6,8 @@
 
 namespace fft
 {
-    typedef int64_t FFT_DATA_TYPE;
-    typedef int64_t FFT_BUTTERFLY_DATA_TYPE;
+    typedef int16_t FFT_DATA_TYPE;
+    typedef int32_t FFT_BUTTERFLY_DATA_TYPE;
 
     const FFT_DATA_TYPE FFT_DATA_TYPE_MAX = INT_FAST16_MAX;
     const FFT_DATA_TYPE FFT_DATA_TYPE_MIN = INT_FAST16_MIN;
